requirements.txt
flask>=2.0.0
nltk>=3.6.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.0
tensorflow-cpu>=2.6.0
keras>=2.6.0
pymupdf>=1.18.0
psycopg2-binary>=2.9.0
gunicorn>=20.1.0
flask-cors==4.0.0